Grind main/menu_favo: FavoDraw icon-loop iconY single-prec frsp wrap (96.11% -> 96.48%)
The icon-draw loop's iconY computation mirrors the bar-draw loop's barY
idiom in the Ghidra ground truth (80162360_FavoDraw, lines 237-238 bar vs
258-259 icon: both compute (int)((float)(halfH/2.0) + (float)y) with the
same single-precision rounding). The bar loop already wraps the sum in
static_cast<float> to match; the icon loop was missing the parallel wrap,
producing a divergent frsp schedule. Adding the wrap makes iconY faithful
to the shared idiom.

FavoDraw misalign 17 -> 12, fuzzy 95.27% -> 96.10%; unit fuzzy 96.11% -> 96.48%.

+++ src/menu_favo.cpp
@@ -196,7 +196,7 @@ void CMenuPcs::FavoDraw()
 		int iconX = static_cast<int>(static_cast<float>(drawEntry->x + drawEntry->w - 0x10));
 		float iconHalfH = static_cast<float>(drawEntry->h) - 32.0f;
 		float iconYf = static_cast<float>(drawEntry->y);
-		int iconY = static_cast<int>(iconHalfH / 2.0 + iconYf);
+		int iconY = static_cast<int>(static_cast<float>(iconHalfH / 2.0 + iconYf));
 		DrawSingleIcon(static_cast<char>(rank->foodId) + 0x14, iconX, iconY, drawEntry->alpha, 1, 1.0f);
 		drawEntry++;
 		rank++;
